Header and positions of calendar configured
